implemented delete function with authorization

+++ controller/reviewController.js
@@ -47,6 +47,13 @@ export function getReviews(req,res){
 export function deleteReview(req,res){
    const email = req.params.email;
 
+   if(req.user==null){
+    res.status(401).json({message:"Please login and try again"});
+    return
+   }
+
+   if(req.user.role=="admin"){
+
     Review.deleteOne
     ({email:email}).then(()=>{
         res.json({message:"Review deleted successfully"});
@@ -54,6 +61,22 @@ export function deleteReview(req,res){
         res.status(500).json
         ({error:"Review deletion failed"});
     });
+      return;
+   }
+    
+   if(req.user.role=="customer"){
+     if(req.user.email==email){
+        Review.deleteOne
+       ({email:email}).then(()=>{
+         res.json({message:"Review deleted successfully"});
+       }).catch(()=>{
+         res.status(500).json
+        ({error:"Review deletion failed"});
+    });
 
+     }else{
+        res.status(403).json({message:"You are not authorized to perform this action"})
+     }
+   } 
 }
